Add files via upload
In this project, I visualize an Argentina Real Estate I found online.
I wrangled and cleaned the data set and also made some statistical calculations. I visualized the relationship between the following:
1. The provinces and approximate prices [Argentine Pesos] using a map to know where the properties are saturated and their corresponding price.
2. The mean house price per square meter by province using a bar chart to know the variation in the house price per square meter to the provinces.
3. The mean house price by province using a bar chart to know the variation in the house price to the provinces.
4. The approximate price Vs. surface covered using a scatter plot.
5. The Buenos Aires approximate price Vs. surface covered using a scatter plot.

I calculated a general Pearson correlation between the surface area covered and the price for the provinces and the Pearson correlation between the surface area and the approximate Price for the Buenos Aires province.

Looking forward to working with you.